[cpp] Use uint32_t to represent field masks (#1537)
In various places (mainly the field_mask() functions) we previously
represented field masks with uint16_t's even though the ObjHeader
allocates 24 bits for them. Now we use uint32_t instead.

I'm motivated to do this now because my upcoming change to move the
object header outside of objects results in one type requiring 17 bits
for its field mask (the field masks get one bit longer).

+++ cpp/frontend_match.h
@@ -34,7 +34,7 @@ class SimpleLexer {
     return ObjHeader::ClassFixed(field_mask(), sizeof(SimpleLexer));
   }
 
-  static constexpr uint16_t field_mask() {
+  static constexpr uint32_t field_mask() {
     return maskbit(offsetof(SimpleLexer, s_));
   }
